integrated webpack for ts bundling
