+++ package.json
@@ -9,11 +9,14 @@
   "dependencies": {
     "@growstocks/stack": "^1.3.0-beta",
     "lit": "^2.0.0-rc.2",
-    "typescript": "^4.3.2"
+    "typescript": "^4.3.2",
+    "webpack": "^5.40.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "ts-loader": "^9.2.3"
+  },
   "scripts": {
-    "build": "tsc",
+    "build": "webpack --config webpack.config.js",
     "build:watch": "tsc --watch"
   },
   "repository": {
