+++ src/js/index.js
@@ -1,0 +1,1 @@
+(()=>{"use strict";var e=document.querySelector(".header--scrollable"),t=document.querySelector(".banner");new IntersectionObserver((function(t,s){t.forEach((function(t){t.isIntersecting?e.classList.remove("scrolled","elevation--z3"):e.classList.add("scrolled","elevation--z3")}))}),{rootMargin:"-100px 0px 0px 0px"}).observe(t);var s=document.querySelector(".search"),l=document.querySelector(".search-button"),i=document.querySelector(".close-button"),o=document.querySelector(".header__text");l.addEventListener("click",(function(){o.classList.add("invisible"),l.classList.add("invisible"),s.classList.add("visible"),i.classList.add("visible")})),i.addEventListener("click",(function(){s.classList.remove("visible"),i.classList.remove("visible"),o.classList.remove("invisible"),l.classList.remove("invisible")}));var r=document.querySelector(".type--file"),n=document.querySelector(".form__file-preview__image"),a=document.querySelector(".form__file-preview__label");r.addEventListener("change",(function(){var e=this.files[0];if(e){var t=new FileReader;n.style.display="block",a.style.display="none",t.addEventListener("load",(function(){console.log(this),n.setAttribute("src",this.result)})),t.readAsDataURL(e)}else n.style.display=null,a.style.display=null}));var c=localStorage.getItem("darkMode"),d=document.querySelector("body"),u=document.querySelector("#dark-mode__toggle");function v(){d.classList.add("dark-mode"),localStorage.setItem("darkMode","enabled"),console.log("Lights off!")}"enabled"===c&&v(),u.addEventListener("click",(function(){"enabled"!==(c=localStorage.getItem("darkMode"))?v():(d.classList.remove("dark-mode"),localStorage.setItem("darkMode",null),console.log("Lights on!"))})),$(document).ready((function(){$(".chip__options__trigger").on("click",(function(){$(this).parent().children(".chip__options").slideToggle("fast");var e=$(this).next();e.hasClass("visible")?($(this).html("menu"),e.removeClass("visible")):($(this).html("close"),e.addClass("visible"))}))}))})();
