+++ src/ts/core.ts
@@ -1,4 +1,3 @@
-"use strict";
 //
 //    Stack by GrowStocks
 //    Made by Matteu#0001
@@ -8,95 +7,102 @@
 //    The Design System for building quick, beautiful, responsive, and flexible user interfaces.
 //
 //
-const header = document.querySelector('.header--scrollable');
-const banner = document.querySelector('.banner');
-const options = {
-    rootMargin: "-100px 0px 0px 0px",
-};
-// @ts-ignore
-const observer = new IntersectionObserver(function (entries, observer) {
-    entries.forEach(entry => {
-        if (!entry.isIntersecting) {
-            header.classList.add('scrolled', 'elevation--z3');
-        }
-        else {
-            header.classList.remove('scrolled', 'elevation--z3');
-        }
-    });
-}, options);
-observer.observe(banner);
-const search = document.querySelector('.search');
-const searchButton = document.querySelector('.search-button');
-const closeButton = document.querySelector('.close-button');
-const headerText = document.querySelector('.header__text');
+import Header from "./structs/Header";
+
+const search: any = document.querySelector('.search')
+const searchButton: any = document.querySelector('.search-button')
+const closeButton: any = document.querySelector('.close-button')
+const headerText: any = document.querySelector('.header__text')
+
 searchButton.addEventListener('click', function () {
-    headerText.classList.add('invisible');
-    searchButton.classList.add('invisible');
-    search.classList.add('visible');
-    closeButton.classList.add('visible');
+
+    headerText.classList.add('invisible')
+    searchButton.classList.add('invisible')
+
+    search.classList.add('visible')
+    closeButton.classList.add('visible')
+
 });
+
 closeButton.addEventListener('click', function () {
-    search.classList.remove('visible');
-    closeButton.classList.remove('visible');
-    headerText.classList.remove('invisible');
-    searchButton.classList.remove('invisible');
+
+    search.classList.remove('visible')
+    closeButton.classList.remove('visible')
+
+    headerText.classList.remove('invisible')
+    searchButton.classList.remove('invisible')
+
 });
+
 // FILE PREVIEW
-const fileInput = document.querySelector('.type--file');
-const previewImg = document.querySelector('.form__file-preview__image');
-const previewText = document.querySelector('.form__file-preview__label');
-fileInput.addEventListener('change', function () {
+const fileInput: any = document.querySelector('.type--file');
+const previewImg: any = document.querySelector('.form__file-preview__image');
+const previewText: any = document.querySelector('.form__file-preview__label');
+
+fileInput.addEventListener('change', function() {
     // @ts-ignore
-    const file = this.files[0];
+    const file: any = this.files[0];
+
     if (!file) {
         previewImg.style.display = null;
         previewText.style.display = null;
-    }
-    else {
+    } else {
         const reader = new FileReader();
+
         previewImg.style.display = "block";
         previewText.style.display = "none";
-        reader.addEventListener("load", function () {
+
+        reader.addEventListener("load", function() {
             console.log(this);
             previewImg.setAttribute('src', this.result);
         });
+
         reader.readAsDataURL(file);
     }
 });
+
 // DARK MODE
-let darkMode = localStorage.getItem("darkMode");
-const theme = document.querySelector('body');
-const toggle = document.querySelector("#dark-mode__toggle");
+let darkMode: any = localStorage.getItem("darkMode");
+const theme: any = document.querySelector('body');
+const toggle: any = document.querySelector("#dark-mode__toggle");
+
 function enable() {
     theme.classList.add('dark-mode');
     localStorage.setItem("darkMode", "enabled");
     console.log('Lights off!');
 }
+
 function disable() {
     theme.classList.remove('dark-mode');
     //@ts-ignore
     localStorage.setItem("darkMode", null);
     console.log('Lights on!');
 }
-if (darkMode === "enabled") {
+
+if(darkMode === "enabled") {
     enable();
 }
-toggle.addEventListener("click", function () {
+
+toggle.addEventListener("click", function() {
+
     darkMode = localStorage.getItem("darkMode");
+
     if (darkMode !== "enabled") {
         enable();
-    }
-    else {
+    } else {
         disable();
     }
 });
+
 // CHIP__OPTIONS
+
 /*const optionsTrigger = document.querySelector('.chip__options__trigger');
 const optionsMenu = document.querySelector('.chip__options');
 
 optionsTrigger.addEventListener('click', function() {
     this.parentElement.firstChild
 });*/
+
 // @ts-ignore
 $(document).ready(function () {
     // @ts-ignore
@@ -109,12 +115,10 @@ $(document).ready(function () {
             // @ts-ignore
             $(this).html("menu");
             option.removeClass("visible");
-        }
-        else {
+        } else {
             // @ts-ignore
             $(this).html("close");
             option.addClass("visible");
         }
     });
 });
-//# sourceMappingURL=_core.js.map
