+++ webpack.config.js
@@ -1,0 +1,22 @@
+const path = require("path");
+
+module.exports = {
+  entry: './src/ts/core.ts',
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/,
+      },
+    ],
+  },
+	mode: "production",
+  resolve: {
+    extensions: ['.tsx', '.ts', '.js'],
+  },
+  output: {
+    filename: 'index.js',
+    path: path.resolve(__dirname, 'src/js'),
+  },
+};
